Make vectorize a subcommand
